test cases
This commit contains all the working test cases till commit f47f610a99b4c686b01773581c0871378a854289 where all the data was coming from the frontend.
The reason I have to use the old commit as the base for this branch is because recently several new commits have been pushed to the main branch which were breaking changes where the data is now coming from the backend and was not even working until yesterday.