fake stats and resume update
Previously, the fake stats was displayed near the footer of the
screen but now it is being diplayed at the top navbar.

Earlier we were using a package to show the multistep form which
was not quite flexible. In this commit we have replaced it with
our own custom multistep form.